Vitest: remove custom dom setup
