+++ vite.config.ts
@@ -2,8 +2,12 @@ import {defineConfig} from 'vitest/config';
 
 export default defineConfig({
 	test: {
-		setupFiles: [
-			'./test/setup-file.js',
-		],
+		environment: 'happy-dom',
+		environmentOptions: {
+			// eslint-disable-next-line @typescript-eslint/naming-convention
+			happyDOM: {
+				url: 'https://github.com',
+			},
+		},
 	},
 });
